Initial refactoring idea
