+++ src/main/java/com/github/loki4j/common/LokiResponse.java
@@ -1,0 +1,11 @@
+package com.github.loki4j.common;
+
+public final class LokiResponse {
+    public int status;
+    public String body;
+    
+    public LokiResponse(int status, String body) {
+        this.status = status;
+        this.body = body;
+    }
+}
