+++ src/main/java/com/github/loki4j/logback/JavaHttpSender.java
@@ -11,10 +11,12 @@ import java.util.concurrent.SynchronousQueue;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 
+import com.github.loki4j.common.LokiResponse;
+
 /**
- * Loki appender that is backed by Java standard {@link java.net.http.HttpClient HttpClient}
+ * Loki sender that is backed by Java standard {@link java.net.http.HttpClient HttpClient}
  */
-public class LokiJavaHttpAppender extends AbstractLoki4jAppender {
+public class JavaHttpSender extends AbstractHttpSender {
 
     private HttpClient client;
     private HttpRequest.Builder requestBuilder;
@@ -22,7 +24,9 @@ public class LokiJavaHttpAppender extends AbstractLoki4jAppender {
     private ExecutorService internalHttpThreadPool;
 
     @Override
-    protected void startHttp(String contentType) {
+    public void start(String contentType) {
+        super.start(contentType);
+
         internalHttpThreadPool = new ThreadPoolExecutor(
             0, Integer.MAX_VALUE,
             getBatchTimeoutMs() * 5, TimeUnit.MILLISECONDS, // expire unused threads after 5 batch intervals
@@ -43,12 +47,14 @@ public class LokiJavaHttpAppender extends AbstractLoki4jAppender {
     }
 
     @Override
-    protected void stopHttp() {
+    public void stop() {
         internalHttpThreadPool.shutdown();
+
+        super.stop();
     }
 
     @Override
-    protected CompletableFuture<LokiResponse> sendAsync(byte[] batch) {
+    public CompletableFuture<LokiResponse> sendAsync(byte[] batch) {
         // Java HttpClient natively supports async API
         // But we have to use its sync API to preserve the ordering of batches
         return CompletableFuture
